create new component wallet info for lixi
