+++ packages/app-lixi/src/components/Wallet/FullWallet.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import WalletInfoComponent from './WalletInfo';
+
+const FullWalletComponent: React.FC = () => {
+  return(
+    <>
+      <WalletInfoComponent />
+      {/* Todo : add history in this page */}
+    </>
+  );
+}
+
+export default FullWalletComponent;
