+++ packages/app-lixi/src/containers/Sidebar/SidebarContent.tsx
@@ -64,6 +64,16 @@ const SidebarContent = ({
               </a>
             </Link>
           </Menu.Item>
+          <Menu.Item>
+            <Link href='/wallet' >
+              <a>
+                <img src='/images/lotus-logo-small.png' alt='lotus' 
+                  style={{width: "20px", marginBottom: "5px", marginRight: "10px"}}
+                />
+                <span>Wallet</span>
+              </a>
+            </Link>
+          </Menu.Item>
         </MenuCustom>
       </div>
     </>
